+++ lib/fog/azurerm/models/compute/availability_set.rb
@@ -11,6 +11,8 @@ module Fog
         attribute :resource_group
         attribute :platform_update_domain_count
         attribute :platform_fault_domain_count
+        attribute :is_managed
+        attribute :sku_name
 
         def self.parse(availability_set)
           hash = {}
@@ -21,6 +23,12 @@ module Fog
           hash['resource_group'] = get_resource_group_from_id(availability_set.id)
           hash['platform_update_domain_count'] = availability_set.platform_update_domain_count
           hash['platform_fault_domain_count'] = availability_set.platform_fault_domain_count
+
+          unless availability_set.sku.nil?
+            hash['sku_name'] = availability_set.sku.name
+            hash['is_managed'] = availability_set.sku.name.eql?(AS_SKU_ALIGNED)
+          end
+
           hash
         end
 
@@ -29,7 +37,7 @@ module Fog
           requires :location
           requires :resource_group
           # need to create the availability set
-          as = service.create_availability_set(availability_set_params(platform_fault_domain_count, platform_update_domain_count))
+          as = service.create_availability_set(availability_set_params(platform_fault_domain_count, platform_update_domain_count, is_managed))
           hash = Fog::Compute::AzureRM::AvailabilitySet.parse(as)
           merge_attributes(hash)
         end
@@ -40,13 +48,14 @@ module Fog
 
         private
 
-        def availability_set_params(platform_fault_domain_count, platform_update_domain_count)
+        def availability_set_params(platform_fault_domain_count, platform_update_domain_count, is_managed)
           {
             name: name,
             location: location,
             resource_group: resource_group,
             platform_fault_domain_count: platform_fault_domain_count,
-            platform_update_domain_count: platform_update_domain_count
+            platform_update_domain_count: platform_update_domain_count,
+            is_managed: is_managed
           }
         end
       end