+++ lib/fog/azurerm/requests/compute/create_availability_set.rb
@@ -9,10 +9,12 @@ module Fog
           resource_group = availability_set_params[:resource_group]
           fault_domain_count = availability_set_params[:platform_fault_domain_count]
           update_domain_count = availability_set_params[:platform_update_domain_count]
+          is_managed = availability_set_params[:is_managed].nil? ? false : availability_set_params[:is_managed]
 
           msg = "Creating Availability Set '#{name}' in #{location} region."
           Fog::Logger.debug msg
-          avail_set_params = get_availability_set_properties(location, fault_domain_count, update_domain_count)
+          avail_set_params = get_availability_set_properties(location, fault_domain_count, update_domain_count, is_managed)
+
           begin
             availability_set = @compute_mgmt_client.availability_sets.create_or_update(resource_group, name, avail_set_params)
           rescue MsRestAzure::AzureOperationError => e
@@ -23,16 +25,31 @@ module Fog
         end
 
         # create the properties object for creating availability sets
-        def get_availability_set_properties(location, fault_domain_count, update_domain_count)
+        def get_availability_set_properties(location, fault_domain_count, update_domain_count, is_managed)
           avail_set = Azure::ARM::Compute::Models::AvailabilitySet.new
-          avail_set.platform_fault_domain_count = fault_domain_count.nil? ? FAULT_DOMAIN_COUNT : fault_domain_count
-          avail_set.platform_update_domain_count = update_domain_count.nil? ? UPDATE_DOMAIN_COUNT : update_domain_count
+
+          if is_managed
+            avail_set.platform_fault_domain_count = fault_domain_count.nil? ? MANAGED_FAULT_DOMAIN_COUNT : fault_domain_count
+            avail_set.platform_update_domain_count = update_domain_count.nil? ? MANAGED_UPDATE_DOMAIN_COUNT : update_domain_count
+          else
+            avail_set.platform_fault_domain_count = fault_domain_count.nil? ? UNMANAGED_FAULT_DOMAIN_COUNT : fault_domain_count
+            avail_set.platform_update_domain_count = update_domain_count.nil? ? UNMANAGED_UPDATE_DOMAIN_COUNT : update_domain_count
+          end
+
           avail_set.virtual_machines = []
           avail_set.statuses = []
           avail_set.location = location
+          avail_set.sku = create_availability_set_sku(is_managed)
           avail_set
         end
+
+        def create_availability_set_sku(is_managed)
+          sku = Azure::ARM::Compute::Models::Sku.new
+          sku.name = is_managed ? AS_SKU_ALIGNED : AS_SKU_CLASSIC
+          sku
+        end
       end
+
       # This class provides the mock implementation for unit tests.
       class Mock
         def create_availability_set(resource_group, name, location)
@@ -43,8 +60,8 @@ module Fog
             'type' => 'Microsoft.Compute/availabilitySets',
             'properties' =>
             {
-              'platformUpdateDomainCount' => FAULT_DOMAIN_COUNT,
-              'platformFaultDomainCount' => FAULT_DOMAIN_COUNT
+              'platformUpdateDomainCount' => UNMANAGED_FAULT_DOMAIN_COUNT,
+              'platformFaultDomainCount' => UNMANAGED_FAULT_DOMAIN_COUNT
             }
           }
         end