+++ test/api_stub/models/compute/availability_set.rb
@@ -3,15 +3,33 @@ module ApiStub
     module Compute
       # Mock class for Availability Set Model
       class AvailabilitySet
-        def self.create_availability_set_response(sdk_compute_client)
+        def self.create_unmanaged_availability_set_response(sdk_compute_client)
           avail_set = {
             'id' => '/subscriptions/########-####-####-####-############/resourceGroups/fog-test-rg/providers/Microsoft.Compute/availabilitySets/fog-test-availability-set',
             'name' => 'fog-test-availability-set',
             'type' => 'Microsoft.Compute/availabilitySets',
             'location' => 'westus',
-            'platformUpdateDomainCount' => UPDATE_DOMAIN_COUNT,
-            'platformFaultDomainCount' => FAULT_DOMAIN_COUNT
+            'platformUpdateDomainCount' => UNMANAGED_UPDATE_DOMAIN_COUNT,
+            'platformFaultDomainCount' => UNMANAGED_FAULT_DOMAIN_COUNT,
+            'sku' => {
+              'name' => 'Classic'
+            }
+          }
+          result_mapper = Azure::ARM::Compute::Models::AvailabilitySet.mapper
+          sdk_compute_client.deserialize(result_mapper, avail_set, 'result.body')
+        end
 
+        def self.create_managed_availability_set_response(sdk_compute_client)
+          avail_set = {
+            'id' => '/subscriptions/########-####-####-####-############/resourceGroups/fog-test-rg/providers/Microsoft.Compute/availabilitySets/fog-test-availability-set',
+            'name' => 'fog-test-availability-set',
+            'type' => 'Microsoft.Compute/availabilitySets',
+            'location' => 'westus',
+            'platformUpdateDomainCount' => MANAGED_UPDATE_DOMAIN_COUNT,
+            'platformFaultDomainCount' => MANAGED_FAULT_DOMAIN_COUNT,
+            'sku' => {
+              'name' => 'Aligned'
+            }
           }
           result_mapper = Azure::ARM::Compute::Models::AvailabilitySet.mapper
           sdk_compute_client.deserialize(result_mapper, avail_set, 'result.body')
@@ -23,9 +41,8 @@ module ApiStub
               'name' => 'fog-test-availability-set',
               'type' => 'Microsoft.Compute/availabilitySets',
               'location' => 'westus',
-              'platformUpdateDomainCount' => UPDATE_DOMAIN_COUNT,
-              'platformFaultDomainCount' => FAULT_DOMAIN_COUNT
-
+              'platformUpdateDomainCount' => UNMANAGED_UPDATE_DOMAIN_COUNT,
+              'platformFaultDomainCount' => UNMANAGED_FAULT_DOMAIN_COUNT
           }
           result_mapper = Azure::ARM::Compute::Models::AvailabilitySet.mapper
           sdk_compute_client.deserialize(result_mapper, avail_set, 'result.body')
@@ -37,9 +54,11 @@ module ApiStub
               'name' => 'fog-test-availability-set',
               'type' => 'Microsoft.Compute/availabilitySets',
               'location' => 'westus',
-              'platformUpdateDomainCount' => UPDATE_DOMAIN_COUNT,
-              'platformFaultDomainCount' => FAULT_DOMAIN_COUNT
-
+              'platformUpdateDomainCount' => UNMANAGED_UPDATE_DOMAIN_COUNT,
+              'platformFaultDomainCount' => UNMANAGED_FAULT_DOMAIN_COUNT,
+              'sku' => {
+                'name' => 'Classic'
+              }
           }
           result_mapper = Azure::ARM::Compute::Models::AvailabilitySet.mapper
           sdk_compute_client.deserialize(result_mapper, avail_set, 'result.body')